+++ packages/platforms/src/main.ts
@@ -1,6 +1,7 @@
 import { EmbedlyPlatformType } from "@embedly/types";
 import { CBC } from "./CBC.ts";
 import { Instagram } from "./Instagram.ts";
+import { Threads } from "./Threads.ts";
 import { TikTok } from "./TikTok.ts";
 import { Twitter } from "./Twitter.ts";
 
@@ -8,5 +9,6 @@ export default {
   [EmbedlyPlatformType.Twitter]: new Twitter(),
   [EmbedlyPlatformType.Instagram]: new Instagram(),
   [EmbedlyPlatformType.TikTok]: new TikTok(),
-  [EmbedlyPlatformType.CBC]: new CBC()
+  [EmbedlyPlatformType.CBC]: new CBC(),
+  [EmbedlyPlatformType.Threads]: new Threads()
 };