+++ packages/builder/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @embedly/builder
 
+## 0.3.0
+
+### Minor Changes
+
+- added threads and fixed a few emoji issues
+
+### Patch Changes
+
+- Updated dependencies
+  - @embedly/parser@0.3.0
+  - @embedly/types@0.3.0
+
 ## 0.2.0
 
 ### Minor Changes