+++ packages/config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @embedly/config
 
+## 0.3.0
+
+### Minor Changes
+
+- added threads and fixed a few emoji issues
+
 ## 0.2.0
 
 ### Minor Changes